docs: added more method signatures
